Removed Pair in the compiler and made more things compile time

+++ ksqldb-execution/src/main/java/io/confluent/ksql/execution/interpreter/terms/BasicTerms.java
@@ -168,7 +168,7 @@ public final class BasicTerms {
 
     @Override
     public SqlType getSqlType() {
-      return SqlTypes.DOUBLE;
+      return SqlTypes.INTEGER;
     }
   }
